backend/requirements.txt
fastapi
uvicorn
requests
pypdf
google-generativeai
openai
anthropic
cryptography
tiktoken
flask
flask-cors
pdfplumber
beautifulsoup4
psycopg2-binary
gunicorn
xhtml2pdf
